Add files via upload
EmployeeDetails class is created with all the logic part along with global variables: name as string data type, age as int data type, salary as float datatype and designation as string datatype. To get user input I have taken Scanner object, create() method is created to take the inputs of the employees. display() method to display the all the details of the required employee, salaryraise() to increment the salary, salary raise for the employee is 1000. main class is created in that object is created for the EmployeeDetails class to fetch the business logic. while loop is taken to repeat the conditions given below, scanner is taken to take the input as choice, if we take choice as 1 it creates the employee details, 2 as choice then it displays the details of the employee, 3 as a choice then it show the message salary is raised, if the choice is greater than equal to 4 then the program should exit.